+++ library/src/main/kotlin/fr/xgouchet/elmyr/StringConstraint.kt
@@ -8,6 +8,7 @@ enum class StringConstraint {
     WORD,
     LIPSUM,
     HEXADECIMAL,
+    NUMERICAL,
     URL,
     URI,
     EMAIL,