+++ library/src/test/kotlin/fr/xgouchet/elmyr/ForgerSpek_Collections.kt
@@ -313,11 +313,12 @@ class ForgerSpek_Collections : Spek({
 
             it("forges a list of whatever") {
                 var i = 0
-                val data = forger.aList { it.anElementFrom(i++) }
+                val data = forger.aList { anElementFrom(i++) }
                 val expectedData = IntArray(data.size) { it }.toTypedArray()
 
                 assertThat(data).containsExactly(*expectedData)
             }
+
         }
 
         context("forging sublist") {